+++ docs/changes.rst
@@ -15,6 +15,7 @@ Released on October 20th 2018.
 * Fixed translation upload API failure in some corner cases.
 * Fixed updating Git configuration in case it was changed manually.
 * Added support for component and project removal.
+* Improved performance for some monolingual translations.
 
 weblate 3.2.1
 -------------